+++ app/helpers/theme_helper.rb
@@ -1,0 +1,9 @@
+module ThemeHelper
+  def current_theme
+    cookies[:theme] || "system"
+  end
+
+  def resolved_theme_class
+    (current_theme == "dark") ? "dark" : ""
+  end
+end
