+++ app/javascript/controllers/dark_mode_controller.js
@@ -1,4 +1,5 @@
 import { Controller } from '@hotwired/stimulus'
+import { getCookie, setCookie } from 'utils/cookies'
 
 export default class extends Controller {
   static targets = ['select']
@@ -8,29 +9,23 @@ export default class extends Controller {
     this.setupSystemThemeListener()
   }
 
+  disconnect () {
+    window.matchMedia('(prefers-color-scheme: dark)').removeEventListener('change', this.onSystemChange.bind(this))
+  }
+
   initializeTheme () {
     const currentTheme = this.getCurrentTheme()
     if (this.hasSelectTarget) {
       this.selectTarget.value = currentTheme
     }
 
-    // Only apply theme if it's not system or if system theme isn't already applied correctly
     if (currentTheme !== 'system') {
       this.applyTheme(currentTheme)
-    } else {
-      // For system theme, verify it's correctly applied
-      const systemPrefersDark = window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches
-      const isDarkApplied = document.documentElement.classList.contains('dark')
-
-      // Only adjust if there's a mismatch
-      if (systemPrefersDark !== isDarkApplied) {
-        this.applyTheme('system')
-      }
     }
   }
 
   getCurrentTheme () {
-    const savedTheme = this.getCookie('theme')
+    const savedTheme = getCookie('theme')
 
     if (savedTheme) {
       return savedTheme
@@ -40,13 +35,13 @@ export default class extends Controller {
   }
 
   setupSystemThemeListener () {
-    if (window.matchMedia) {
-      const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)')
-      mediaQuery.addEventListener('change', () => {
-        if (this.getCurrentTheme() === 'system') {
-          this.applyTheme('system')
-        }
-      })
+    if (!window.matchMedia) return
+    window.matchMedia('(prefers-color-scheme: dark)').addEventListener('change', this.onSystemChange.bind(this))
+  }
+
+  onSystemChange () {
+    if (this.getCurrentTheme() === 'system') {
+      this.applyTheme('system')
     }
   }
 
@@ -54,7 +49,7 @@ export default class extends Controller {
     const theme = this.selectTarget.value
 
     this.applyTheme(theme)
-    this.setCookie('theme', theme)
+    setCookie('theme', theme)
   }
 
   applyTheme (mode) {
@@ -69,22 +64,5 @@ export default class extends Controller {
     }
 
     document.documentElement.classList.toggle('dark', isDark)
-  }
-
-  getCookie (name) {
-    const nameEQ = name + '='
-    const ca = document.cookie.split(';')
-    for (let i = 0; i < ca.length; i++) {
-      let c = ca[i]
-      while (c.charAt(0) === ' ') c = c.substring(1, c.length)
-      if (c.indexOf(nameEQ) === 0) return c.substring(nameEQ.length, c.length)
-    }
-    return null
-  }
-
-  setCookie (name, value) {
-    const date = new Date()
-    date.setTime(date.getTime() + (365 * 24 * 60 * 60 * 1000))
-    document.cookie = `${name}=${value}; expires=${date.toUTCString()}; path=/; SameSite=Lax`
   }
 }
