+++ config/importmap.rb
@@ -8,6 +8,7 @@ pin_all_from "app/javascript/controllers", under: "controllers"
 pin "@stimulus-components/dropdown", to: "@stimulus-components--dropdown.js" # @3.0.0
 pin "stimulus-use" # @0.52.2
 pin_all_from "app/javascript/custom", under: "custom"
+pin_all_from "app/javascript/utils", under: "utils"
 pin "trix"
 pin "@rails/actiontext", to: "actiontext.esm.js"
 pin "@stimulus-components/password-visibility", to: "@stimulus-components--password-visibility.js" # @3.0.0
